+++ app/FoodHeroTests/Search/Domain/RestaurantSearchWithStubTests.m
@@ -38,9 +38,9 @@
     _search = [[RestaurantSearch alloc] initWithRestaurantRepository:_restaurantRepository locationService:[(id <ApplicationAssembly>) [TyphoonComponents factory] locationService]];
 }
 
-- (NSMutableArray *)restaurants:(int)nrRestaurants {
+- (NSMutableArray *)restaurants:(NSInteger)nrRestaurants {
     NSMutableArray *restaurants = [NSMutableArray new];
-    for (int i = 0; i < nrRestaurants; i++) {
+    for (NSInteger i = 0; i < nrRestaurants; i++) {
         [restaurants addObject:[[RestaurantBuilder alloc] build]];
     }
     return restaurants;
